+++ static/js/lib/api/ckeditor.js
@@ -1,0 +1,4 @@
+// @flow
+import { fetchWithAuthFailure } from "./fetch_auth"
+
+export const getCKEditorJWT = () => fetchWithAuthFailure("/api/v0/ckeditor/")
