+++ static/js/lib/api/fetch_auth.js
@@ -8,8 +8,8 @@ import {
 } from "redux-hammock/django_csrf_fetch"
 import qs from "query-string"
 
-import { AUTH_REQUIRED_URL, LOGIN_URL } from "./url"
-import { isNotAuthenticatedErrorType } from "../util/rest"
+import { AUTH_REQUIRED_URL, LOGIN_URL } from "../url"
+import { isNotAuthenticatedErrorType } from "../../util/rest"
 
 const redirectAndReject = async () => {
   // redirect to the authenticating app
