+++ static/js/reducers/auth.js
@@ -1,6 +1,6 @@
 // @flow
 import R from "ramda"
-import * as api from "../lib/api"
+import * as api from "../lib/api/api"
 import { INITIAL_STATE } from "redux-hammock/constants"
 import { deriveVerbFuncs } from "redux-hammock/hammock"
