+++ static/js/reducers/frontpage.js
@@ -1,7 +1,7 @@
 // @flow
 import { GET, INITIAL_STATE } from "redux-hammock/constants"
 
-import * as api from "../lib/api"
+import * as frontpageAPI from "../lib/api/frontpage"
 import { mapPostListResponse } from "../lib/posts"
 
 import type { PostListPaginationParams } from "../flow/discussionTypes"
@@ -16,6 +16,7 @@ export const frontPageEndpoint = {
       postIds:    []
     }
   },
-  getFunc:           (params: PostListPaginationParams) => api.getFrontpage(params),
+  getFunc: (params: PostListPaginationParams) =>
+    frontpageAPI.getFrontpage(params),
   getSuccessHandler: mapPostListResponse
 }
