+++ static/js/reducers/post_removed.js
@@ -1,10 +1,11 @@
 // @flow
-import * as api from "../lib/api"
+import * as postAPI from "../lib/api/posts"
 import { PATCH, INITIAL_STATE } from "redux-hammock/constants"
 
 export const postRemovedEndpoint = {
-  name:         "postRemoved",
-  verbs:        [PATCH],
-  patchFunc:    (id: string, removed: boolean) => api.updateRemoved(id, removed),
+  name:      "postRemoved",
+  verbs:     [PATCH],
+  patchFunc: (id: string, removed: boolean) =>
+    postAPI.updateRemoved(id, removed),
   initialState: { ...INITIAL_STATE }
 }
