+++ static/js/reducers/post_upvotes.js
@@ -1,10 +1,11 @@
 // @flow
-import * as api from "../lib/api"
+import * as postAPI from "../lib/api/posts"
 import { PATCH, INITIAL_STATE } from "redux-hammock/constants"
 
 export const postUpvotesEndpoint = {
-  name:         "postUpvotes",
-  verbs:        [PATCH],
-  patchFunc:    (id: string, upvoted: boolean) => api.updateUpvote(id, upvoted),
+  name:      "postUpvotes",
+  verbs:     [PATCH],
+  patchFunc: (id: string, upvoted: boolean) =>
+    postAPI.updateUpvote(id, upvoted),
   initialState: { ...INITIAL_STATE }
 }
