+++ static/js/reducers/profile_image.js
@@ -1,6 +1,6 @@
 // @flow
 import { PATCH, INITIAL_STATE } from "redux-hammock/constants"
-import * as api from "../lib/api"
+import * as api from "../lib/api/api"
 
 export const profileImageEndpoint = {
   name:         "profileImage",
