+++ static/js/reducers/subscribedChannels.js
@@ -2,13 +2,13 @@
 import { GET, INITIAL_STATE } from "redux-hammock/constants"
 
 import type { Channel } from "../flow/discussionTypes"
-import * as api from "../lib/api"
+import * as channelAPI from "../lib/api/channels"
 
 export const subscribedChannelsEndpoint = {
   name:              "subscribedChannels",
   verbs:             [GET],
   initialState:      { ...INITIAL_STATE, data: [] },
-  getFunc:           () => api.getChannels(),
+  getFunc:           () => channelAPI.getChannels(),
   getSuccessHandler: (payload: Array<Channel>): any =>
     payload.map(channel => channel.name)
 }
